requirements.txt
python-dotenv>=1.0.1,<2.0.0
fastapi>=0.115.0,<1.0.0
uvicorn>=0.32.0,<1.0.0
aiohttp>=3.10.0,<4.0.0
